feat(apollo,look&feel): update loading button and move loader to apollo
