+++ apps/look-and-feel-stories-css/src/Button.stories.ts
@@ -1,4 +1,4 @@
-import "@axa-fr/design-system-look-and-feel-css/dist/Button/Button.scss";
+import "@axa-fr/design-system-apollo-css/dist/Button/ButtonLF.scss";
 import type { Meta, StoryObj } from "@storybook/html";
 import { render } from "./ButtonRender";
