+++ client/apollo/css/src/Button/ButtonCommon.scss
@@ -31,6 +31,10 @@
   &:disabled,
   &[aria-disabled="true"] {
     pointer-events: none;
+
+    & > [aria-live] {
+      margin: unset;
+    }
   }
 
   & > svg {
