+++ client/apollo/css/src/Loader/Loader.scss
@@ -1,0 +1,39 @@
+.af-loader {
+  &__container {
+    display: flex;
+    container-type: inline-size;
+    margin: auto;
+    border: 1px solid var(--color-gray-200);
+    border-top: 1px solid var(--color-axa);
+    border-radius: 50%;
+    flex-direction: column;
+    align-items: center;
+    justify-content: center;
+    animation: spin 2s linear infinite;
+
+    &--blue {
+      border: solid color-mix(in srgb, var(--color-axa), transparent 80%);
+      border-top: solid var(--color-axa);
+    }
+
+    &--gray {
+      border: solid color-mix(in srgb, var(--color-gray), transparent 80%);
+      border-top: solid var(--color-gray);
+    }
+
+    &--white {
+      border: solid color-mix(in srgb, var(--color-white), transparent 80%);
+      border-top: solid var(--color-white);
+    }
+  }
+}
+
+@keyframes spin {
+  0% {
+    transform: rotate(0deg);
+  }
+
+  100% {
+    transform: rotate(360deg);
+  }
+}
