+++ client/apollo/css/src/apollo.scss
@@ -1,1 +1,2 @@
 @use "./Button/ButtonApollo";
+@use "./Loader/Loader";
