+++ client/apollo/css/src/look-and-feel.scss
@@ -1,1 +1,2 @@
 @use "./Button/ButtonLF";
+@use "./Loader/Loader";
