+++ client/apollo/react/src/index.ts
@@ -4,3 +4,4 @@ import "@fontsource/source-sans-pro";
 export { Button } from "./Button/ButtonApollo";
 export { buttonVariants, type ButtonVariants } from "./Button/ButtonCommon";
 export { Svg } from "./Svg/Svg";
+export { Loader } from "./Loader/Loader";
