+++ client/apollo/react/src/indexLF.ts
@@ -4,3 +4,4 @@ import "@fontsource/source-sans-pro";
 export { Button } from "./Button/ButtonLF";
 export { buttonVariants, type ButtonVariants } from "./Button/ButtonCommon";
 export { Svg } from "./Svg/Svg";
+export { Loader } from "./Loader/Loader";
