--- client/look-and-feel/css/src/Button/Button.scss
@@ -1,154 +1,0 @@
-.af-btn-client {
-  display: flex;
-  min-height: 3.5rem;
-  padding: 0.8rem 1.2rem;
-  border: 0;
-  border-radius: 8px;
-  align-items: center;
-  justify-content: center;
-  gap: 1rem;
-  font-family: var(--font-family-base);
-  font-weight: 600;
-  text-decoration: none;
-  color: var(--color-white);
-  background-color: var(--color-axa);
-  transition-duration: var(--transition-duration);
-  transition-property: width, height, border, color, background-color, outline,
-    box-shadow;
-  transition-timing-function: linear;
-  cursor: pointer;
-  user-select: none;
-
-  &:hover {
-    text-decoration: none;
-    color: var(--color-white);
-    background-color: var(--color-blue-1);
-  }
-
-  &:active,
-  &:focus {
-    color: var(--color-white);
-    background-color: var(--color-btn-light);
-    outline: none;
-  }
-
-  &:focus-visible {
-    outline: 2px solid var(--color-blue-3);
-    outline-offset: 3px;
-  }
-
-  & > svg {
-    aspect-ratio: 1;
-    fill: currentcolor;
-  }
-
-  &--ghost {
-    min-width: auto;
-    min-height: auto;
-    padding: 0;
-    gap: 0.25rem;
-    color: var(--color-axa);
-    background-color: transparent;
-
-    &:hover {
-      color: var(--color-blue-1);
-      background-color: transparent;
-    }
-
-    &:focus,
-    &:active {
-      color: var(--color-btn-light);
-      background-color: transparent;
-    }
-  }
-
-  &--primary-business {
-    background-color: var(--color-btn-business);
-
-    &:hover {
-      background-color: var(--color-btn-business-dark);
-    }
-
-    &:focus {
-      background-color: var(--color-btn-business-light);
-    }
-  }
-
-  &--tertiary {
-    color: var(--color-axa);
-    background-color: var(--color-btn-tertiary-bg);
-
-    &:hover {
-      color: var(--color-axa);
-      background-color: var(--color-btn-tertiary-bg-darker);
-    }
-
-    &:focus,
-    &:active {
-      color: var(--color-axa);
-      background-color: var(--color-btn-tertiary-bg-lighter);
-    }
-  }
-
-  &--secondary {
-    color: var(--color-axa);
-    background-color: var(--color-white);
-    box-shadow: 0 0 0 1px var(--color-axa) inset;
-
-    &:hover,
-    &:focus,
-    &:active {
-      color: var(--color-blue-1);
-      background-color: var(--color-white);
-      box-shadow: 0 0 0 2px var(--color-blue-1) inset;
-    }
-
-    &:active {
-      background-color: var(--color-blue-2);
-    }
-  }
-
-  &--secondary-error {
-    color: var(--color-red-700);
-    background-color: var(--color-white);
-    box-shadow: 0 0 0 1px var(--color-red-700) inset;
-
-    &:hover,
-    &:focus,
-    &:active {
-      color: var(--color-btn-error-text);
-      background-color: var(--color-white);
-      box-shadow: 0 0 0 2px var(--color-btn-error-text) inset;
-    }
-
-    &:active {
-      background-color: var(--color-btn-error-bg);
-    }
-  }
-
-  &__wrapper {
-    display: flex;
-    align-items: center;
-    justify-content: center;
-    line-height: 1;
-  }
-
-  &:disabled,
-  &[aria-disabled="true"] {
-    color: var(--color-btn-disabled-text);
-    background-color: var(--color-gray-300);
-    box-shadow: none;
-    cursor: not-allowed;
-
-    &:hover,
-    &:focus {
-      border-color: transparent;
-      color: var(--color-btn-disabled-text);
-      background-color: var(--color-gray-300);
-    }
-
-    & > [aria-live] {
-      margin: unset;
-    }
-  }
-}
