+++ client/look-and-feel/css/src/look-and-feel.scss
@@ -3,7 +3,6 @@
 @use "Grid/DebugGrid";
 @use "Accordion/Accordion";
 @use "AccordionCore/AccordionCore";
-@use "Button/Button";
 @use "Card/Card";
 @use "Divider/Divider";
 @use "Layout/Footer/Footer";
@@ -22,7 +21,6 @@
 @use "Pagination/Pagination";
 @use "Pagination/PaginationDesktop";
 @use "Pagination/PaginationMobile";
-@use "Loader/Loader";
 @use "./List/List";
 @use "List/ClickList/ClickList";
 @use "./List/ContentItemDuo/ContentItemDuo";
