--- client/look-and-feel/react/src/Loader/index.ts
@@ -1,1 +1,0 @@
-export { Loader } from "./Loader";
