+++ vitest.workspace.ts
@@ -1,3 +1,7 @@
 import { defineWorkspace } from "vitest/config";
 
-export default defineWorkspace(["slash/react", "client/look-and-feel/react"]);
+export default defineWorkspace([
+  "slash/react",
+  "client/look-and-feel/react",
+  "client/apollo/react",
+]);
